Extended image attributes

These functions extract and replace medical image attributes that go beyond the core metadata associated with the NIfTI file formats.

imageAttributes(x)

imageAttributes(x) <- value

Arguments

x

A niftiImage object.

value

A list of new image attributes to replace any existing ones.

Value

A list of image attributes, or a modified object with these changed. These are essentially all attributes except those used for basic niftiImage objects.

Details

The DICOM (Digital Imaging and Communications in Medicine) format, and BIDS (Brain Imaging Data Structure), which extends NIfTI, can both encapsulate copious amounts of metadata about a scan and the patient. This metadata can be useful for more advanced or research-focussed post-processing methods, and standard R attributes are a natural place to store it. The imageAttributes function returns a list of just these extended attributes, if they exist, ignoring other attributes used by the package. The replacement form allows this metadata to be modified or removed. These functions currently only act on objects inheriting from the niftiImage class.

References

More information about metadata captured by the BIDS format can be found at https://bids.neuroimaging.io or in the paper cited below.

K.J. Gorgolewski, T. Auer, V.D. Calhoun, et al. The brain imaging data structure, a format for organizing and describing outputs of neuroimaging experiments (2016). Scientific Data 3:160044. doi:10.1038/sdata.2016.44 .

See also

Author

Jon Clayden <code@clayden.org>

Examples

path <- system.file("extdata", "example.nii.gz", package="RNifti")
image <- readNifti(path)
imageAttributes(image)
#> named list()