Internal images

An internal image is a simple R object with a few attributes including a pointer to an internal C structure, which contains the full image data. They are used in the package for efficiency, but can be converted to a normal R array using the as.array method. Attributes of these objects should not be changed.

# S3 method for internalImage
dim(x)

# S3 method for internalImage
dim(x) <- value

# S3 method for internalImage
as.array(x, ...)

# S3 method for internalImage
[(x, i, j, ..., drop = TRUE)

# S3 method for internalImage
[(x, i, j, ...) <- value

Arguments

x
An "internalImage" object.

value
Not used. Changing the dimensions of (or data in) an internal image is invalid, and will produce an error. Convert to an array first.

...
Additional parameters to methods. Only used for additional indices.

i, j
Index vectors. May be missing, which indicates that the whole of the relevant dimension should be obtained.

drop
If TRUE (the default), unitary indices in the result will be dropped. This mirrors the behaviour of standard array indexing.

Author

Jon Clayden <code@clayden.org>