Check the format version of a file

This function identifies the likely NIfTI format variant used by one or more files on disk.

niftiVersion(file)

Arguments

file
A character vector of file names.

Value

A vector of integers, of the same length as file. Each element will be 0 for ANALYZE format (the precursor to NIfTI-1), 1 for NIfTI-1 (which is now most common), 2 for NIfTI-2, or -1 if the file doesn't exist or doesn't look plausible in any of these formats.

Note

NIfTI-2 format, mostly a variant of NIfTI-1 with wider datatypes used for many fields, is not currently supported for reading, but it is detected by this function.

See also

Examples

path <- system.file("extdata", "example.nii.gz", package="RNifti")
niftiVersion(path)       # 1

#> /tmp/RtmpZyHYpi/temp_libpath7e285a37f74a/RNifti/extdata/example.nii.gz 
#>                                                                      1